Fix deleting coupon issue

+++ src/main/java/com/lpdm/msorder/service/impl/CouponServiceImpl.java
@@ -52,12 +52,18 @@ public class CouponServiceImpl implements CouponService {
      * Delete a {@link Coupon}
      * @param coupon The {@link Coupon} object to be deleted
      * @return True if the {@link Coupon} has been deleted, otherwise false
+     * @throws CouponNotFoundException Thrown if no {@link Coupon} was found in the database
      */
     @Override
-    public boolean deleteCoupon(Coupon coupon) {
+    public boolean deleteCoupon(Coupon coupon) throws CouponNotFoundException {
 
-        couponRepository.delete(coupon);
         Optional<Coupon> optCoupon = couponRepository.findById(coupon.getId());
+        if(!optCoupon.isPresent()) throw new CouponNotFoundException();
+
+        coupon.setCode(optCoupon.get().getCode());
+        couponRepository.delete(coupon);
+
+        optCoupon = couponRepository.findById(coupon.getId());
 
         return !optCoupon.isPresent();
     }
